moved package back to users. Refactored ACE login out into it's own class so make it more re-usable
