+++ bundle/src/main/java/com/adobe/acs/commons/users/impl/ProtectedSystemUsers.java
@@ -1,4 +1,4 @@
-package com.adobe.acs.commons.principals.impl;
+package com.adobe.acs.commons.users.impl;
 
 import org.apache.commons.lang.ArrayUtils;
